+++ src/main/java/org/voucherProject/voucherProject/repository/voucher/VoucherRepository.java
@@ -1,4 +1,4 @@
-package org.voucherProject.voucherProject.repository;
+package org.voucherProject.voucherProject.repository.voucher;
 
 import org.voucherProject.voucherProject.entity.voucher.Voucher;
