+++ src/test/java/org/voucherProject/voucherProject/controller/VoucherControllerImplTest.java
@@ -3,6 +3,7 @@ package org.voucherProject.voucherProject.controller;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.voucherProject.voucherProject.controller.voucher.VoucherController;
 import org.voucherProject.voucherProject.entity.voucher.Voucher;
 import org.voucherProject.voucherProject.entity.voucher.VoucherType;
