+++ src/test/java/org/voucherProject/voucherProject/repository/VoucherRepositoryFileImplTest.java
@@ -7,6 +7,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.voucherProject.voucherProject.entity.voucher.FixedAmountVoucher;
 import org.voucherProject.voucherProject.entity.voucher.PercentDiscountVoucher;
 import org.voucherProject.voucherProject.entity.voucher.Voucher;
+import org.voucherProject.voucherProject.repository.voucher.VoucherRepositoryFileImpl;
 
 import java.io.IOException;
 import java.util.List;
